+++ packages/babel/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@linaria/babel",
+  "name": "@linaria/babel-preset",
   "version": "3.0.0-beta.0",
   "publishConfig": {
     "access": "public"
