+++ packages/core/__tests__/detect-core-js.test.js
@@ -14,7 +14,7 @@ const waitForProcess = async (process) => {
 
 it('Ensures that package do not include core-js dependency after build', async () => {
   // eslint-disable-next-line import/no-extraneous-dependencies
-  const packageJSON = require('@linaria/babel/package.json');
+  const packageJSON = require('@linaria/babel-preset/package.json');
   const buildScript = packageJSON.scripts['build:lib'];
 
   const proc = cp.exec(buildScript, {
