+++ packages/extractor/__tests__/extractor.test.ts
@@ -1,4 +1,4 @@
-import { run } from '@linaria/babel/__utils__/strategy-tester';
+import { run } from '@linaria/babel-preset/__utils__/strategy-tester';
 
 describe('extractor', () => {
   run(__dirname, require('../src').default);
