+++ packages/extractor/package.json
@@ -41,7 +41,7 @@
   "dependencies": {
     "@babel/generator": ">=7",
     "@babel/plugin-transform-runtime": ">=7",
-    "@linaria/babel": "^3.0.0-beta.0",
+    "@linaria/babel-preset": "^3.0.0-beta.0",
     "@linaria/preeval": "^3.0.0-beta.0"
   },
   "peerDependencies": {
