+++ packages/extractor/tsconfig.json
@@ -1,8 +1,5 @@
 {
   "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "paths": {},
-    "rootDir": "src/"
-  },
+  "compilerOptions": { "paths": {}, "rootDir": "src/" },
   "references": [{ "path": "../babel" }, { "path": "../preeval" }]
 }
