+++ packages/preeval/package.json
@@ -38,7 +38,7 @@
     "watch": "yarn build --watch"
   },
   "dependencies": {
-    "@linaria/babel": "^3.0.0-beta.0"
+    "@linaria/babel-preset": "^3.0.0-beta.0"
   },
   "peerDependencies": {
     "@babel/core": ">=7"
