+++ packages/rollup/package.json
@@ -43,7 +43,7 @@
     "rollup": "1.20.0||^2.0.0"
   },
   "dependencies": {
-    "@linaria/babel": "^3.0.0-beta.0",
+    "@linaria/babel-preset": "^3.0.0-beta.0",
     "@rollup/pluginutils": "^4.1.0"
   },
   "peerDependencies": {
