+++ packages/shaker/package.json
@@ -44,7 +44,7 @@
   "dependencies": {
     "@babel/generator": ">=7",
     "@babel/plugin-transform-runtime": ">=7",
-    "@linaria/babel": "^3.0.0-beta.0",
+    "@linaria/babel-preset": "^3.0.0-beta.0",
     "@linaria/logger": "^3.0.0-beta.0",
     "@linaria/preeval": "^3.0.0-beta.0",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
