+++ packages/shaker/src/identifierHandlers.ts
@@ -1,6 +1,6 @@
 import { types as t } from '@babel/core';
 import type { Aliases, Identifier, Node, VisitorKeys } from '@babel/types';
-import { peek } from '@linaria/babel';
+import { peek } from '@linaria/babel-preset';
 import GraphBuilderState from './GraphBuilderState';
 import type { IdentifierHandlerType, NodeType } from './types';
 import { identifierHandlers as core } from './langs/core';
