+++ packages/shaker/src/langs/core.ts
@@ -23,7 +23,7 @@ import type {
   WhileStatement,
 } from '@babel/types';
 
-import { peek } from '@linaria/babel';
+import { peek } from '@linaria/babel-preset';
 import type { IdentifierHandlers, Visitors } from '../types';
 import GraphBuilderState from '../GraphBuilderState';
 import ScopeManager from '../scope';
