+++ packages/stylelint/package.json
@@ -36,7 +36,7 @@
     "watch": "yarn build --watch"
   },
   "dependencies": {
-    "@linaria/babel": "^3.0.0-beta.0",
+    "@linaria/babel-preset": "^3.0.0-beta.0",
     "strip-ansi": "^5.2.0"
   }
 }
