+++ packages/stylelint/src/preprocessor.ts
@@ -1,6 +1,6 @@
 import stripAnsi from 'strip-ansi';
-import { transform } from '@linaria/babel';
-import type { Replacement } from '@linaria/babel';
+import { transform } from '@linaria/babel-preset';
+import type { Replacement } from '@linaria/babel-preset';
 
 type Errors = {
   [key: string]:
