+++ packages/webpack5-loader/package.json
@@ -47,7 +47,7 @@
     "webpack": "^5.6.0"
   },
   "dependencies": {
-    "@linaria/babel": "3.0.0-beta.0",
+    "@linaria/babel-preset": "3.0.0-beta.0",
     "@linaria/logger": "3.0.0-beta.0",
     "cosmiconfig": "^5.1.0",
     "enhanced-resolve": "^5.3.1",
